requirements.txt
pymongo==4.6.2
python-dotenv==1.0.1
pydantic==2.6.4
pytest==8.0.2
pylint==3.1.0
black==24.3.0
radon==6.0.1
bandit==1.7.8
pytest-cov==4.1.0
numpy==1.26.4
scikit-learn==1.4.1
